# Conflicts:
#	capacity_parent/capacity_common/src/main/java/com/zhidejiaoyu/common/mapper/LearnExtendMapper.java
#	capacity_parent/capacity_common/src/main/java/com/zhidejiaoyu/common/mapper/LearnExtendMapper.xml
#	capacity_parent/capacity_common/src/main/java/com/zhidejiaoyu/common/mapper/LearnNewMapper.java
#	capacity_parent/capacity_common/src/main/java/com/zhidejiaoyu/common/mapper/LearnNewMapper.xml
#	capacity_parent/capacity_common/src/main/java/com/zhidejiaoyu/common/mapper/UnitVocabularyNewMapper.java

+++ capacity_parent/capacity_common/src/main/java/com/zhidejiaoyu/common/pojo/StudentStudyPlanNew.java
@@ -54,7 +54,7 @@ public class StudentStudyPlanNew extends Model<StudentStudyPlanNew> {
     /**
      * 学习流程表id（存放难、易模块的流程id）
      */
-    private Integer flowId;
+    private Long flowId;
     /**
      * 难易类型：1：简单类型；2：难类型
      */
